Device#startTest: returns nil or FBTestManager
Previously, this method return FBTestManager instance on success and
and NSError subclass on failure.  The returned NSError was not
consumed by the callers, instead the NSError passed by reference
was consumed which means that all the extra context provided
by the returned XCTestBootstrapError was never displayed to the user.

Now #startTest returns nil on failure and correct sets the NSError
reference with the XCTestBootstrapError (for extra context) and
the callers branch on nil FBTestManager.