Simulators: use 'instruments identifier' for alias matching
Previously we checked only the simulator.name which is not unique.

The tests were failing because the default simulator was iPhone 8
(11.4) but the find-simulator-by-name algorithm in iOSDeviceManager was
matching on the first 'iPhone 8' which happened to be 'iPhone 8 (11.0)'.

'instruments identifier' refers to the string that we used to pass
to instruments when running under UIAutomation.

Co-authored-by: Tikhon Polyakov <polyakrecords@gmail.com>